Add support of parsing exception stream
"Correctly" (this stream is not documented properly, so it's
implementation-dependant) parse exception stream in order to
extract exception chain records.

This is pretty important, since the information about a crash (
which then generated the crashdump) is stored in the exception stream.
This is the entry point to analyze automatically crash dumps.